Support checkpointing for server side initalization of nnunet plans
Modified initialize of NnunetServer to still sample a client and get properties if nnunet_plans is in config in order to get other properties (num_segmentation_heads, num_input_channels and enable_deep_supervision. These are required to instantiate the model on the server side which is a prerequisite for checkpointing.